backend/requirements.txt
fastapi
uvicorn
pokerkit==0.4.3
psycopg2-binary
python-dotenv
